+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## FIWARE Big Bang v0.30.0-next
 
+-   Fix nifi.properties not found (#295)
+
 ## FIWARE Big Bang v0.30.0 - 04 August, 2023
 
 -   Fix issue that Keyrock fails to create application with openid (#292)